Actually fix the PHPCS error
